Changed .nstd from 0 to 3
The time series dat is passed to MatLab's POLYFIT routine to obtain a
polynomial fit PP to dat, and the difference dat-PP is compared to the
original dat. If the standard deviation of (dat-PP) is less than that of
dat by a factor of NSTD, the time series is assumed to contain a
significant trend (output will be 0), else not (output will be 1).

With nstd=0 this will always detect a trend (and therefore output 0).